Fix missing translations and moved helper functions
