+++ app/javascript/components/vm-disks/helper.js
@@ -1,30 +1,22 @@
 import { headerData, rowData } from '../miq-data-table/helper';
-import { toHumanSize } from '../workers-form/helpers';
-
-const numberToHumanSize = (bytes) => {
-  const human = toHumanSize(bytes);
-  if (!human) {
-    return '';
-  }
-  return human.replace(/^[\d.]+/, (n) => Math.round(parseFloat(n)));
-};
+import { toHumanSize } from '../../helpers/size';
 
 const diskDisplayName = ({ device_type: deviceType, controller_type: controllerType, location }) => {
   switch (deviceType) {
-    case 'cdrom-raw':     return `CD-ROM (IDE ${location})`;
-    case 'atapi-cdrom':   return `ATAPI CD-ROM (IDE ${location})`;
-    case 'cdrom-image':   return `CD-ROM Image (IDE ${location})`;
-    case 'ide':           return `Hard Disk (IDE ${location})`;
+    case 'cdrom-raw':     return sprintf(__('CD-ROM (IDE %s)'), location);
+    case 'atapi-cdrom':   return sprintf(__('ATAPI CD-ROM (IDE %s)'), location);
+    case 'cdrom-image':   return sprintf(__('CD-ROM Image (IDE %s)'), location);
+    case 'ide':           return sprintf(__('Hard Disk (IDE %s)'), location);
     case 'scsi':
-    case 'scsi-hardDisk': return `Hard Disk (SCSI ${location})`;
-    case 'scsi-passthru': return `Generic SCSI (${location})`;
-    case 'floppy':        return `Floppy Drive (SIO ${location})`;
+    case 'scsi-hardDisk': return sprintf(__('Hard Disk (SCSI %s)'), location);
+    case 'scsi-passthru': return sprintf(__('Generic SCSI (%s)'), location);
+    case 'floppy':        return sprintf(__('Floppy Drive (SIO %s)'), location);
     case 'disk':
       if (controllerType && controllerType.startsWith('ide')) {
-        return `Hard Disk (IDE ${location})`;
+        return sprintf(__('Hard Disk (IDE %s)'), location);
       }
       if (controllerType && controllerType.startsWith('scsi')) {
-        return `Hard Disk (SCSI ${location})`;
+        return sprintf(__('Hard Disk (SCSI %s)'), location);
       }
       return `${controllerType} ${location}`;
     default: return `${controllerType} ${location}`;
@@ -40,8 +32,8 @@ export const tableData = (disks) => {
       { text: disk.disk_type || '' },
       { text: disk.mode || '' },
       { text: disk.partitions_aligned || __('Unknown') },
-      { text: disk.size ? numberToHumanSize(disk.size) : '' },
-      { text: disk.size_on_disk ? numberToHumanSize(disk.size_on_disk) : '' },
+      { text: disk.size ? toHumanSize(disk.size) : '' },
+      { text: disk.size_on_disk ? toHumanSize(disk.size_on_disk) : '' },
       { text: (disk.size && disk.size_on_disk) ? ((disk.size_on_disk / disk.size) * 100).toFixed(1) : '' },
     ],
   }));
