+++ app/javascript/components/workers-form/workers-form.jsx
@@ -5,9 +5,8 @@ import { get } from 'lodash';
 import addSchema from './workers.schema';
 import MiqFormRenderer from '../../forms/data-driven-form';
 import { API } from '../../http_api';
-import {
-  toBytes, parseWorker, buildPatch, toRubyMethod,
-} from './helpers';
+import { parseWorker, buildPatch, toRubyMethod } from './helpers';
+import { toBytes } from '../../helpers/size';
 import './workers.style.scss';
 
 const WorkersForm = ({ server: { id, name }, product, zone }) => {
